+++ src/containers/ProfileDetails/components/CaptivePortal/index.js
@@ -1,13 +1,25 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useMemo } from 'react';
 import PropTypes from 'prop-types';
-import { Card, Form, Input, Radio, Select, Upload, Alert, Collapse, message, List } from 'antd';
-import { QuestionCircleFilled } from '@ant-design/icons';
-
+import {
+  Card,
+  Form,
+  Input,
+  Radio,
+  Select,
+  Upload,
+  Alert,
+  Collapse,
+  message,
+  List,
+  Table,
+} from 'antd';
+import { QuestionCircleFilled, FormOutlined, DeleteOutlined } from '@ant-design/icons';
 import Button from 'components/Button';
+import Modal from 'components/Modal';
 import Tooltip from 'components/Tooltip';
+import FormModal from './components/FormModal';
 
 import styles from '../index.module.scss';
-import Users from './components/Users';
 
 const { Item } = Form;
 const { Option } = Select;
@@ -52,11 +64,64 @@ const CaptivePortalForm = ({ details, form, fileUpload, radiusProfiles }) => {
   const [bgFileList, setBgFileList] = useState(
     (details.backgroundFile && [formatFile(details.backgroundFile)]) || []
   );
-
   const [whitelist, setWhitelist] = useState(details.walledGardenAllowlist || []);
   const [whitelistSearch, setWhitelistSearch] = useState();
   const [whitelistValidation, setWhitelistValidation] = useState({});
 
+  const [userList, setUserList] = useState(details.userList || []);
+
+  const [addUserModal, setAddUserModal] = useState(false);
+  const [editUserModal, setEditUserModal] = useState(false);
+  const [deleteUserModal, setDeleteUserModal] = useState(false);
+  const [activeUser, setActiveUser] = useState({});
+
+  const columns = [
+    {
+      title: 'Username',
+      dataIndex: 'username',
+      width: 250,
+    },
+    {
+      title: 'First Name',
+      dataIndex: ['userDetails', 'firstName'],
+      width: 150,
+    },
+    {
+      title: 'Last Name',
+      dataIndex: ['userDetails', 'lastName'],
+    },
+    {
+      width: 64,
+      render: (_, record) => (
+        <Button
+          className={styles.InfoButton}
+          title={`edit-${record.username}`}
+          type="primary"
+          icon={<FormOutlined />}
+          onClick={() => {
+            setEditUserModal(true);
+            setActiveUser({ ...record });
+          }}
+        />
+      ),
+    },
+
+    {
+      width: 64,
+      render: (_, record) => (
+        <Button
+          className={styles.InfoButton}
+          title={`delete-${record.username}`}
+          type="danger"
+          icon={<DeleteOutlined />}
+          onClick={() => {
+            setDeleteUserModal(true);
+            setActiveUser({ ...record });
+          }}
+        />
+      ),
+    },
+  ];
   const disableExternalSplashChange = () => {
     form.setFieldsValue({
       authenticationType: 'guest',
@@ -254,16 +319,39 @@ const CaptivePortalForm = ({ details, form, fileUpload, radiusProfiles }) => {
 
   const handleDeleteWhitelist = item => setWhitelist(whitelist.filter(i => i !== item));
 
-  const handleAddUser = newUser => {
-    details.userList.push(newUser);
+  const handleAddUser = ({ username, password, firstName, lastName }) => {
+    const newUser = {
+      username,
+      password,
+      userDetails: {
+        firstName,
+        lastName,
+      },
+    };
+    setUserList([...userList, newUser]);
+    setAddUserModal(false);
   };
-  const handleUpdateUser = (index, newUser) => {
-    details.userList.splice(index, 1, newUser);
+  const handleUpdateUser = ({ username, password, firstName, lastName }) => {
+    const newUser = {
+      username,
+      password,
+      userDetails: {
+        firstName,
+        lastName,
+      },
+    };
+    setUserList(userList.map(user => (user.username === activeUser.username ? newUser : user)));
+    setEditUserModal(false);
   };
 
-  const handleDeleteUser = index => {
-    details.userList.splice(index);
+  const handleDeleteUser = () => {
+    setUserList(userList.filter(user => user.username !== activeUser.username));
+    setDeleteUserModal(false);
   };
+
+  const usedUserNames = useMemo(() => {
+    return userList.map(i => i.username);
+  }, [userList]);
 
   useEffect(() => {
     form.setFieldsValue({
@@ -291,6 +379,12 @@ const CaptivePortalForm = ({ details, form, fileUpload, radiusProfiles }) => {
       walledGardenAllowlist: whitelist,
     });
   }, [whitelist]);
+
+  useEffect(() => {
+    form.setFieldsValue({
+      userList,
+    });
+  }, [userList]);
 
   return (
     <div className={styles.ProfilePage}>
@@ -372,14 +466,46 @@ const CaptivePortalForm = ({ details, form, fileUpload, radiusProfiles }) => {
       </Card>
 
       {authentication === 'username' && (
-        <Users
-          userList={details.userList}
-          handleAddUser={handleAddUser}
-          handleUpdateUser={handleUpdateUser}
-          handleDeleteUser={handleDeleteUser}
-        />
+        <>
+          <Modal
+            onCancel={() => setDeleteUserModal(false)}
+            onSuccess={handleDeleteUser}
+            visible={deleteUserModal}
+            title="Are you sure?"
+            buttonText="Delete"
+            buttonType="danger"
+            content={
+              <p>
+                Are you sure you want to delete the user: <strong>{activeUser.username}</strong>
+              </p>
+            }
+          />
+          <FormModal
+            onCancel={() => setEditUserModal(false)}
+            visible={editUserModal}
+            onSubmit={handleUpdateUser}
+            title="Edit User"
+            username={activeUser.username}
+            firstName={activeUser?.userDetails?.firstName}
+            lastName={activeUser?.userDetails?.lastName}
+            usedUserNames={usedUserNames}
+          />
+          <FormModal
+            onCancel={() => setAddUserModal(false)}
+            visible={addUserModal}
+            onSubmit={handleAddUser}
+            title="Add User"
+            usedUserNames={usedUserNames}
+          />
+          <Card
+            title="User List"
+            extra={<Button onClick={() => setAddUserModal(true)}> Add User</Button>}
+          >
+            <Table rowKey="username" columns={columns} dataSource={userList} pagination={false} />
+          </Card>
+          <Item name="userList" noStyle hidden />
+        </>
       )}
-
       {authentication === 'radius' && (
         <Card title="RADIUS">
           <Item
@@ -535,7 +661,6 @@ const CaptivePortalForm = ({ details, form, fileUpload, radiusProfiles }) => {
           </Item>
         </Panel>
       </Collapse>
-
       <Collapse expandIconPosition="right">
         <Panel header="Splash Page Images" forceRender>
           <Item label="Configure">
@@ -602,7 +727,6 @@ const CaptivePortalForm = ({ details, form, fileUpload, radiusProfiles }) => {
           )}
         </Panel>
       </Collapse>
-
       <Collapse expandIconPosition="right">
         <Panel header="Allow List" forceRender>
           <Item
